[clib] Fix some size_t related compiler warnings

+++ include/cantera/clib/ct.h
@@ -102,9 +102,9 @@ extern "C" {
     CANTERA_CAPI int thermo_setState_Psat(int n, double p, double x);
     CANTERA_CAPI int thermo_setState_Tsat(int n, double t, double x);
 
-    CANTERA_CAPI size_t kin_newFromXML(int mxml, int iphase,
-                                       int neighbor1, int neighbor2, int neighbor3,
-                                       int neighbor4);
+    CANTERA_CAPI int kin_newFromXML(int mxml, int iphase,
+                                    int neighbor1, int neighbor2, int neighbor3,
+                                    int neighbor4);
     CANTERA_CAPI int kin_del(int n);
     CANTERA_CAPI size_t kin_nSpecies(int n);
     CANTERA_CAPI size_t kin_nReactions(int n);
@@ -137,7 +137,7 @@ extern "C" {
     CANTERA_CAPI int kin_advanceCoverages(int n, double tstep);
     CANTERA_CAPI size_t kin_phase(int n, size_t i);
 
-    CANTERA_CAPI size_t trans_new(const char* model, int th, int loglevel);
+    CANTERA_CAPI int trans_new(const char* model, int th, int loglevel);
     CANTERA_CAPI int trans_del(int n);
     CANTERA_CAPI double trans_viscosity(int n);
     CANTERA_CAPI double trans_electricalConductivity(int n);
